heroku_backend/requirements.txt
cloudpickle==2.0.0
dask==2.30.0
dataprep==0.4.0
distlib==0.3.3
Flask 
flatbuffers==2.0
gast==0.3.3
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz 
grpcio 
gunicorn==20.1.0
h5py
idna==3.3
importlib-metadata 
jsonify==0.5
jsonpath-ng==1.5.3
jsonschema 
Keras
Keras-Applications
Keras-Preprocessing
matplotlib 
matplotlib-inline 
multidict==5.2.0
numpy==1.21.2
oauthlib==3.1.1
olefile
packaging==21.0
pickleshare 
protobuf==3.18.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow-cpu
tensorflow-addons 
tensorflow-estimator
termcolor==1.1.0 